Add "RegressError" to the "regress" module
By exposing the exception as a module attribute, we ensure that
callers can use that name in `except` clauses.

Without this addition, `regress.RegressError` can still be raised, but
it has a type which can't be referred to by name.

pyo3 docs use `PyModule.add(...)` in a formulation similar to, but
slightly different from, `add_class` usage.