Add support for passing flags as strings
resolves #51

There is no support for `Flags` structs as provided by the `regress`
crate. But a string can be passed down and is converted to Flags by
the `with_flags` constructor.
Exposing flags as a struct/enum adds significant complexity deemed
not worth the additional code at this time.
(See #51 for more details.)

A python test verifies that the alternate constructor pattern works
and provides support for the `"u"` flag.