Remove running pip-audit from CI.
pip has had a recent CVE, and as a library (and not
an app) it is difficult to run pip-audit in a way that
has value but is segregated from pip-audit's own deps
such that we don't encounter this kind of false positive.

Downstream applications using this library can (and should)
be running pip-audit themselves.